Database (#26)
* Add flask-sqlalchemy

* Initialise the app with Sqlalchemy

* Prepare models for Sqlalchemy

* Add Flask Migrate

* Intialise app with Migrate

* Add Sqlalchemy URl configurations

* Add psycopg2

* Add Initial data Migration

* Modify Model & create Migration

* Modify token verification and generation methods

* Modify User Endpoints to use database

* Modify data Models

* Ignore files from tracking

* Integrate the database for the Book Blueprint

The commit also adds the routes to add Book sections"

* Reafactor User model tests

* Add database to book & borrow Blueprints

*  Finish database implemtation

+++ app/book/views.py
@@ -1,9 +1,28 @@
-from flask_restful import Resource, reqparse
+from flask import session
+from flask_restful import Resource, reqparse, abort
 
 from app.data_repo.book_repo import get_all_books, create_book, abort_if_book_exists, \
     get_book_by_id, abort_if_book_not_found, update_book, delete_book
 
+from app.data_repo.user_repo import get_user_by_id
+
 from app.utils.data_validators import string_validator
+
+
+def check_active_session():
+    if 'userID' not in session:
+        msg = "Kindly Login first: Forbidden Action"
+        link = "/api/v1/auth/login"
+        abort(401, message=msg, login_link=link)
+
+
+def check_if_admin(action):
+    """Check that current logged in user is an admin"""
+    user = get_user_by_id(session['userID'])
+    if user.is_admin:
+        return
+    msg = f"Your not authorised to {action}"
+    abort(401, message=msg)
 
 
 class BookResource(Resource):
@@ -13,41 +32,61 @@ class BookResource(Resource):
 
     def __init__(self):
 
-        self.book_parser = reqparse.RequestParser()
+        self.book_update_parser = reqparse.RequestParser()
 
-        self.book_parser.add_argument('name', type=string_validator, required=True,
+        self.book_update_parser.add_argument('name', type=string_validator, required=True,
                                       location='json')
 
-        self.book_parser.add_argument('description', type=string_validator, required=True,
+        self.book_update_parser.add_argument('description', type=string_validator, required=True,
                                       location='json')
 
-        self.book_parser.add_argument('section', type=string_validator, required=True,
-                                      location='json')
+        self.book_update_parser.add_argument('section', type=int, required=True,
+                                             help="Book Section is a Number & cannot ne blank",
+                                             location='json')
 
-        self.book_parser.add_argument('quantity', type=int, required=True,
-                                      help='Book quantity is a Number cannot be blank ',
-                                      location='json')
-
-    def put(self, bookId):
-        abort_if_book_not_found(bookId)
-        book_args = self.book_parser.parse_args()
-
-        book_name = book_args['name']
-        book_desc = book_args['description']
-        book_section = book_args['section']
-        book_qty = book_args['quantity']
-
-        book = update_book(bookId, name=book_name, description=book_desc,
-                                  section=book_section, quantity=book_qty)
-
-        response = {'message': "Book:%s was updated" % bookId, 'data': book}
-        return response, 200
+        self.book_update_parser.add_argument('quantity', type=int, required=True,
+                                             help='Book quantity is a Number & cannot be blank ',
+                                             location='json')
 
     def get(self, bookId):
+        check_active_session()
         abort_if_book_not_found(bookId)
-        return get_book_by_id(bookId)
+        book = get_book_by_id(bookId)
+        return {
+            "name": book.name,
+            "description": book.description,
+            "section": book.section.name,
+            "quantity": book.quantity
+        }, 200
+
+    def put(self, bookId):
+        check_active_session()
+
+        check_if_admin("update a book")
+
+        abort_if_book_not_found(bookId)
+        book_args = self.book_update_parser.parse_args()
+
+        name = book_args['name']
+        description = book_args['description']
+        section = book_args['section']
+        quantity = book_args['quantity']
+
+        book = update_book(bookId, name=name, description=description,
+                                  section=section, quantity=quantity)
+
+        return {
+            'message': "Book:%s was updated" % bookId,
+            'book_name': book.name,
+            'book_decription': book.description,
+            'book_section': book.section.name,
+            'book_copies': book.quantity
+        }, 200
 
     def delete(self, bookId):
+        check_active_session()
+        check_if_admin("delete a book")
+
         abort_if_book_not_found(bookId)
         delete_book(bookId)
 
@@ -71,26 +110,41 @@ class BooksResource(Resource):
         self.book_parser.add_argument('description', type=string_validator, required=True,
                                       location='json')
 
-        self.book_parser.add_argument('section', type=string_validator, required=True,
+        self.book_parser.add_argument('section', type=int, required=True,
+                                      help="Book Section is a Number & cannot be blank",
                                       location='json')
 
         self.book_parser.add_argument('quantity', type=int, required=True,
-                                      help='Book quantity is a Number cannot be blank ',
+                                      help='Book quantity is a Number cannot be blank',
                                       location='json')
 
     def get(self):
-        return get_all_books()
+        check_active_session()
+        books = get_all_books()
+        if not books:
+            return {
+                "message": "There are no books at the moment"
+            }
+        book_list = {}
+        for book in books:
+            book_list[book.id] = book.__repr__()
+
+        return book_list, 200
 
     def post(self):
+        check_active_session()
+        check_if_admin("create a book")
         book_args = self.book_parser.parse_args()
 
-        book_name = book_args['name']
-        book_desc = book_args['description']
-        book_section = book_args['section']
-        book_qty = book_args['quantity']
+        name = book_args['name']
+        description = book_args['description']
+        section = book_args['section']
+        quantity = book_args['quantity']
 
-        abort_if_book_exists(book_name)
+        abort_if_book_exists(name)
 
-        new_book = create_book(book_name, book_desc, book_section, book_qty)
+        book = create_book(name, description, section, quantity)
 
-        return new_book.__dict__, 201
+        return {
+            "message": 'Book Created - {}'.format(book.name)
+        }, 201
